refactor(components): clean up code and update components package structure
- Removed unused imports and components
- Reorganized component folder structure
