+++ app/src/main/java/com/dyshuk/android/upworkredesigncompose/ui/components/buttons/FilledDefaultButton.kt
@@ -1,6 +1,5 @@
-package com.dyshuk.android.upworkredesigncompose.ui.components
+package com.dyshuk.android.upworkredesigncompose.ui.components.buttons
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.shape.RoundedCornerShape
