+++ app/src/main/java/com/dyshuk/android/upworkredesigncompose/ui/components/indicators/ArcProgressBar.kt
@@ -1,4 +1,4 @@
-package com.dyshuk.android.upworkredesigncompose.ui.components
+package com.dyshuk.android.upworkredesigncompose.ui.components.indicators
 
 import androidx.annotation.DrawableRes
 import androidx.compose.animation.core.Animatable
