+++ app/src/main/java/com/dyshuk/android/upworkredesigncompose/ui/components/indicators/StarRating.kt
@@ -1,4 +1,4 @@
-package com.dyshuk.android.upworkredesigncompose.ui.components
+package com.dyshuk.android.upworkredesigncompose.ui.components.indicators
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Row
