+++ app/src/main/java/com/dyshuk/android/upworkredesigncompose/ui/components/text/JobTags.kt
@@ -1,4 +1,4 @@
-package com.dyshuk.android.upworkredesigncompose.ui.components
+package com.dyshuk.android.upworkredesigncompose.ui.components.text
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
